requirements.txt
numpy>=1.19.2
scikit-learn>=1.0.2
pandas>=1.0.0
matplotlib>=3.3.0
Pillow>=8.0.0
pygame>=2.0.0 